wav2lip/requirements.txt
# Easy-Wav2Lip + YOLO11 Requirements
# RTX 3050 + WSL2 最適化版

# Core Dependencies (from original Easy-Wav2Lip)
basicsr==1.4.2
batch-face==1.4.0
# dlib==19.24.2  # 一時的にコメントアウト（ビルド問題回避）
facexlib==0.3.0
gdown==4.7.1
gfpgan==1.3.8
imageio-ffmpeg==0.4.9
importlib-metadata==6.8.0
ipython==8.16.1
librosa==0.10.1
moviepy==1.0.3
numpy==1.26.1
opencv-python==4.8.1.78
scipy==1.11.3

# PyTorch (CUDA 12.1 compatible)
--extra-index-url https://download.pytorch.org/whl/cu121
torch==2.1.0
torchaudio==2.1.0
torchvision==0.16.0

# YOLO11 and Object Detection
ultralytics>=8.0.0
onnx>=1.15.0
onnxruntime-gpu>=1.16.0

# TensorRT Optimization
nvidia-tensorrt
torch-tensorrt
pycuda>=2022.1

# Face Detection (RetinaFace for dynamic resolution)
retinaface-pytorch>=0.0.1

# Additional Dependencies
pillow>=9.0.0
tqdm>=4.65.0
pyyaml>=6.0
easydict>=1.10
argparse>=1.4.0
configparser>=5.3.0

# Face Detection Enhancement
# retinaface-pytorch>=0.0.1  # 一時的にコメントアウト（依存関係問題回避）
insightface>=0.7.3

# Audio Processing
soundfile>=0.12.1
librosa>=0.10.1
# pyaudio>=0.2.11  # 一時的にコメントアウト（ビルド問題回避）

# Video Processing
ffmpeg-python>=0.2.0
imageio>=2.31.0

# CLI Enhancement
click>=8.1.0
rich>=13.0.0
colorama>=0.4.6

# Logging and Monitoring
loguru>=0.7.0
wandb>=0.15.0  # オプション：トレーニング監視用

# FastAPI and Web Server Dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
aiofiles>=23.0.0
starlette>=0.27.0

# Additional System Dependencies
psutil>=5.9.0
requests>=2.31.0

# Development Tools
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0